Put ResumeSelectorComponent in its own component in prep for merge of wizards
