+++ frontend/src/components/SubmitView.js
@@ -14,10 +14,6 @@ import Button from "react-bootstrap/Button";
 import { GoLightBulb, GoZap } from 'react-icons/go';
 
 import "bootstrap/dist/css/bootstrap.css";
-import 'react-bootstrap-typeahead/css/Typeahead.css';
-import 'react-bootstrap-typeahead/css/Typeahead-bs4.css';
-
-import { Typeahead } from 'react-bootstrap-typeahead';
 
 import AceEditor from "react-ace";
 import "ace-builds/src-noconflict/mode-groovy";
@@ -27,10 +23,7 @@ import "ace-builds/src-noconflict/keybinding-sublime";
 
 import { exampleScript, exampleConfig } from "../examples.js"
 
-import { parseStatus, BADGE_STYLES } from "./Widgets.js"
-import { format, formatRelative } from 'date-fns/fp'
-
-const now = new Date();
+import ResumeSelectorComponent from "./ResumeSelectorComponent.js"
 
 const EditorComponent = ({editorRef, value, onChangeHandler, isDarkMode, ...props}) => (
     <AceEditor
@@ -163,38 +156,13 @@ function SubmitView(props) {
             <div className="workflow-detail-well mt-4 p-4" >
             <Row>
             <Col>
-                { resumeSelection 
-                    ? <span>Resume from: {resumeSelection}</span>
-                    : <Typeahead
-                        id="resume-box-selector"
-                        dropup={true}
-                        minLength={0}
-                        placeholder="Select prior run for resume..."
-                        options={resumeData}
-                        isLoading={resumeDataIsLoading}
-                        bsSize="large"
-                        clearButton={true}
-                        onChange={(i) => {
-                            const val = i[0] ? i[0].fargateTaskArn : null;
-                            setResumeSelection(val)
-                        }}
-                        labelKey={(i) => i.nextflowRunName || "unknown"}
-                        renderMenuItemChildren={(option, props, index) => {
-                            const status = parseStatus(option.runnertaskstatus, option.nextflowlastevent);
-                            const date_string = formatRelative(now, new Date(option.fargateCreatedAt)).capFirstLetter()
-                            return (
-                                <div className='searchresult'>
-                                <div className='key'>{option.nextflowRunName || "unknown"}</div>
-                                    <div>
-                                         <span style={{fontSize: 12, color: "#999", paddingRight: 10}}>{date_string}</span>
-                                         <span className={"text-" + BADGE_STYLES[status]} style={{fontSize: 12, fontWeight: "bold"}}>
-                                            {status}
-                                        </span>
-                                    </div>
-                                </div>);
-                        }}
-                        />
-                }
+               <ResumeSelectorComponent 
+                    resumeData={resumeData}
+                    resumeDataIsLoading={resumeDataIsLoading}
+                    resumeSelection={resumeSelection}
+                    setResumeSelection={setResumeSelection}
+                    dropup={true}
+                />
             </Col><Col>
                 <div style={{textAlign: "right"}}>
                     <Button size="lg" onClick={handleSubmit}>Run Workflow <GoZap /></Button>
