+++ frontend/src/components/WizardView.js
@@ -17,16 +17,9 @@ import Button from "react-bootstrap/Button";
 import { GoLightBulb, GoZap } from 'react-icons/go';
 
 import "bootstrap/dist/css/bootstrap.css";
-import 'react-bootstrap-typeahead/css/Typeahead.css';
-import 'react-bootstrap-typeahead/css/Typeahead-bs4.css';
 
-import { Typeahead } from 'react-bootstrap-typeahead';
-
-import { parseStatus, BADGE_STYLES } from "./Widgets.js"
 import UploadEditor from "./UploadEditor.js"
-import { format, formatRelative } from 'date-fns/fp'
-
-const now = new Date();
+import ResumeSelectorComponent from "./ResumeSelectorComponent.js"
 
 const handleError = (response) => {
     if (!response.ok) { throw response }
@@ -120,41 +113,14 @@ function WizardView(props) {
             <Form.Group as={Row} controlId="formResumeSelection">
               <Form.Label column sm={3}>Resume from:</Form.Label>
               <Col sm={9}>
-                    { resumeSelection 
-                    ? <span>
-                        {resumeSelection}
-                        <span style={{float: "right"}}>
-                            <Button onClick={() => setResumeSelection(null)} variant='outline-secondary'>Clear</Button></span>
-                      </span>
-                    : <Typeahead
-                        id="resume-box-selector"
-                        dropup={true}
-                        minLength={0}
-                        placeholder="Select prior run for resume..."
-                        options={resumeData}
-                        isLoading={resumeDataIsLoading}
-                        clearButton={true}
-                        onChange={(i) => {
-                            const val = i[0] ? i[0].fargateTaskArn : null;
-                            setResumeSelection(val)
-                        }}
-                        labelKey={(i) => i.nextflowRunName || "unknown"}
-                        renderMenuItemChildren={(option, props, index) => {
-                            const status = parseStatus(option.runnertaskstatus, option.nextflowlastevent);
-                            const date_string = formatRelative(now, new Date(option.fargateCreatedAt)).capFirstLetter()
-                            return (
-                                <div className='searchresult'>
-                                <div className='key'>{option.nextflowRunName || "unknown"}</div>
-                                    <div>
-                                         <span style={{fontSize: 12, color: "#999", paddingRight: 10}}>{date_string}</span>
-                                         <span className={"text-" + BADGE_STYLES[status]} style={{fontSize: 12, fontWeight: "bold"}}>
-                                            {status}
-                                        </span>
-                                    </div>
-                                </div>);
-                        }}
-                        />
-                }
+                <ResumeSelectorComponent 
+                    resumeData={resumeData}
+                    resumeDataIsLoading={resumeDataIsLoading}
+                    resumeSelection={resumeSelection}
+                    setResumeSelection={setResumeSelection}
+                    dropup={true}
+                    bsSize="medium"
+                />
               </Col>
             </Form.Group>
             <Form.Group as={Row} controlId="formUrl">
